requirements.txt
requests
Pillow
discord.py